measurement setting for all observations
