+++ projects/ngx-charts-on-fhir/src/lib/fhir-mappers/observation/component-observation-mapper.service.ts
@@ -6,6 +6,7 @@ import { DataLayer } from '../../data-layer/data-layer';
 import { Mapper } from '../multi-mapper.service';
 import { ChartAnnotation, isDefined } from '../../utils';
 import { TIME_SCALE_OPTIONS, LINEAR_SCALE_OPTIONS, ANNOTATION_OPTIONS } from '../fhir-mapper-options';
+import { getMeasurementSettingSuffix } from './simple-observation-mapper.service';
 
 /** Required properties for mapping an Observation with [ComponentObservationMapper] */
 export type ComponentObservation = {
@@ -58,7 +59,7 @@ export class ComponentObservationMapper implements Mapper<ComponentObservation> 
       datasets: resource.component
         .sort((a, b) => a.code.text.localeCompare(b.code.text))
         .map((component) => ({
-          label: component.code.text,
+          label: component.code.text + getMeasurementSettingSuffix(resource),
           yAxisID: scaleName,
           data: [
             {
